backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-dotenv==1.0.1
selenium==4.21.0
webdriver-manager==4.0.1
pytest==8.2.0
httpx==0.27.0
pytest-asyncio==0.23.7
anyio[trio]==4.4.0
bcrypt==5.0.0
python-jose[cryptography]==3.3.0
email-validator==2.1.1
pytesseract==0.3.13
Pillow==10.3.0
pytest-mock==3.14.0
freezegun==1.5.1
beautifulsoup4==4.12.3
lxml==5.2.1
